sem_pipeline/requirements.txt
openai
pandas
python-dotenv
sentence-transformers
thefuzz
rapidfuzz
openpyxl
